Add more infrastructure, a little more UI
At this point, I think I want to scrap the tabbed UI and go with a
multi-pane interface. Left side is the account list, right side is the
drill-down for a particular account which shows all the public keys with
copy and edit buttons. If the user hits edit, open a dialog with the
GenerateKeysPage in it.